pyproject.toml
[tool.poetry]
name = "ol_concourse_github_issues"
version = "0.1.0"
description = "concourse.io resource for manipulating Github Issues"
authors = ["Chris Patti <feoh@feoh.org>"]
license = "MIT"
readme = "README.md"
packages = [{include = "ol_concourse_github_issues"}]

[tool.poetry.dependencies]
python = "^3.11"
concoursetools = "^0.7.1"
PyGithub = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pre-commit = "^3.4.0"
mypy = "^1.6.1"
ruff = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.11"
follow_imports = "normal"
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true
ignore_missing_imports = true
# for strict mypy: (this is the tricky one :-))
# disallow_untyped_defs = true
